Create advanced search menu using sliders and radio buttons. Style advanced search menu. Add new complementary google font - Oswald.
