+++ src/components/common/AdvancedSearchMenu.js
@@ -6,6 +6,9 @@ import {
   SettingOutlined,
   UserOutlined
 } from "@ant-design/icons";
+import DecimalSlider from "./DecimalSlider";
+import GraduatedSlider from "./GraduatedSlider";
+import RadioButtons from "./RadioButtons";
 
 const { SubMenu } = Menu;
 
@@ -14,7 +17,7 @@ class AdvancedSearchMenu extends React.Component {
   rootSubmenuKeys = ["sub1", "sub2", "sub3", "sub4", "sub5"];
 
   state = {
-    openKeys: ["sub1"]
+    openKeys: []
   };
 
   onOpenChange = openKeys => {
@@ -36,22 +39,30 @@ class AdvancedSearchMenu extends React.Component {
         mode="inline"
         openKeys={this.state.openKeys}
         onOpenChange={this.onOpenChange}
-        style={{ width: 256 }}
         className="advanced-search-menu"
       >
-        <SubMenu key="sub1" icon={<UserOutlined />} title="Population">
-          <Menu.Item key="1">Option 1</Menu.Item>
-          <Menu.Item key="2">Option 2</Menu.Item>
-          <Menu.Item key="3">Option 3</Menu.Item>
-          <Menu.Item key="4">Option 4</Menu.Item>
+        <SubMenu
+          key="sub1"
+          icon={<UserOutlined />}
+          title="Population"
+          className="advanced-search-submenu"
+        >
+          <p className="advanced-search-option-title">Choose City Size</p>
+          <RadioButtons />
         </SubMenu>
         <SubMenu key="sub2" icon={<AppstoreOutlined />} title="Weather">
-          <Menu.Item key="5">Option 5</Menu.Item>
-          <Menu.Item key="6">Option 6</Menu.Item>
-          <SubMenu key="sub3" title="Submenu">
-            <Menu.Item key="7">Option 7</Menu.Item>
-            <Menu.Item key="8">Option 8</Menu.Item>
-          </SubMenu>
+          <p className="advanced-search-option-title">
+            Choose Average Summer Temperature (°F)
+          </p>
+          <DecimalSlider min={-20} max={100} />
+          <p className="advanced-search-option-title">
+            Choose Average UV Index (0-10)
+          </p>
+          <DecimalSlider min={0} max={10} />
+          <p className="advanced-search-option-title">
+            Choose Average Cloud Cover (%)
+          </p>
+          <DecimalSlider min={0} max={100} />
         </SubMenu>
         <SubMenu key="sub3" icon={<SettingOutlined />} title="Job Industry">
           <Menu.Item key="9">Option 9</Menu.Item>
@@ -60,10 +71,10 @@ class AdvancedSearchMenu extends React.Component {
           <Menu.Item key="12">Option 12</Menu.Item>
         </SubMenu>
         <SubMenu key="sub4" icon={<SettingOutlined />} title="Housing">
-          <Menu.Item key="9">Option 9</Menu.Item>
-          <Menu.Item key="10">Option 10</Menu.Item>
-          <Menu.Item key="11">Option 11</Menu.Item>
-          <Menu.Item key="12">Option 12</Menu.Item>
+          <p className="advanced-search-option-title">
+            Choose Average 1BR Home Prices
+          </p>
+          <GraduatedSlider />
         </SubMenu>
         <SubMenu key="sub5" icon={<SettingOutlined />} title="COVID-19">
           <Menu.Item key="9">Option 9</Menu.Item>
