+++ src/components/pages/AdvancedSearch/AdvancedSearchContainer.js
@@ -15,39 +15,51 @@ const Container = styled.div`
 const MenuWrapper = styled.div`
   background-color: #70c783;
   min-height: 100vh;
-  width: 300px;
+  width: 450px;
   padding: 1rem;
 
   display: flex;
   flex-direction: column;
   align-items: center;
-  justify-content: space-between;
+  /* justify-content: space-between; */
 `;
 const ResultsWrapper = styled.div`
   display: flex;
   flex-direction: column;
   align-items: center;
   margin: 0 auto;
-  width: 100%;
+  width: 60%;
 `;
 
 const AdvancedSearch = () => {
-  const { cities, setIsComparing, addCity, getData } = useContext(CityContext);
+  const {
+    cities,
+    setIsComparing,
+    addCity,
+    getData,
+    weather,
+    suggestions
+  } = useContext(CityContext);
   const history = useHistory();
 
   return (
     <Container>
       <MenuWrapper>
         <Link id="menuLogoWrapper">
-          <DriftlyLogo id="menuLogo" onClick={() => history.push("/")} />
+          <DriftlyLogo id="menuLogo" onClick={() => history.push("/home")} />
         </Link>
         <h3 id="advanced-search-title">Refine your search</h3>
         <AdvancedSearchMenu id="advanced-search-menu" />
+        <div id="get-results-button" role="button">
+          Get Results
+        </div>
       </MenuWrapper>
       <ResultsWrapper>
         <Header />
-        <div>Showing 20 of </div>
-        <div className="container">
+        <div className="results-title">
+          {}Showing 20 of {suggestions.length} results
+        </div>
+        <div className="card-container">
           {cities.slice(30).map((city, index) => (
             <CityCard
               key={index}
@@ -58,6 +70,7 @@ const AdvancedSearch = () => {
               addCity={addCity}
               getData={getData}
               cities={cities}
+              weather={weather}
             />
           ))}
         </div>
