+++ src/components/pages/Home/RenderHomePage.js
@@ -74,7 +74,7 @@ function RenderHomePage() {
           </div>
         )}
 
-        <div className="container">
+        <div className="card-container">
           {cities.map((city, index) => (
             <CityCard
               key={index}
