+++ src/index.css
@@ -14,7 +14,7 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
-.container {
+.card-container {
   width: 90%;
   margin: 0 auto;
   display: flex;
@@ -24,6 +24,16 @@ code {
 .home-page-container {
   display: flex;
   flex-direction: column;
+}
+.results-title {
+  width: 90%;
+  margin: 0 auto;
+  padding: 0 1rem;
+
+  font-family: 'Oswald', sans-serif;
+  font-weight: 400;
+  font-size: 1.5rem;
+  color: #656665;
 }
 
 /* HEADER SECTION */
@@ -195,7 +205,7 @@ code {
   align-items: center;
   flex-direction: column;
   border: 2px solid gold;
-  border-radius: 3px;
+  border-radius: 6px;
 }
 .card-list {
   width: 100%;
@@ -264,7 +274,7 @@ code {
 /* DATA VISUALIZATION PAGE */
 #menuLogoWrapper {
   top: 3%;
-  width: 80%;
+  width: 60%;
 }
 #menuLogo {
   width: 100%;
@@ -404,25 +414,67 @@ code {
 /* ADVANCED SEARCH MENU */
 #advanced-search-title {
   position: absolute;
-  top: 9rem;
+  top: 11.5rem;
   font-size: 2.5rem;
   font-weight: 700;
   text-align: left;
   color: #ffffff;
-  font-family: Amatic SC;
+  font-family: 'Amatic SC', 'sans serif';
 }
 .advanced-search-menu {
   position: absolute;
-  /* width: 100%; */
-  top: 12rem;
-  border-radius: 5px;
-  font-family: Amatic SC;
-  font-size: 1.5rem;
+  top: 15rem;
+  border-radius: 6px;
+  font-family: 'Amatic SC', 'sans serif';
+  font-size: 1.8rem;
   font-weight: 700;
+  width: 400px;
+  color: #656665;
 }
-/* .sub-menu-title {
+.advanced-search-option-title {
+  font-size: 1.2rem;
+  font-family: 'Oswald', 'sans-serif';
+  padding: 2rem 1rem 0 1rem;
+}
+.advanced-search-input {
+  font-family: 'Oswald', 'sans serif';
+}
+.ant-row {
+  padding: 0 1rem;
+}
+.ant-slider-with-marks {
+  margin: 34px 1.5rem;
+  font-family: 'Oswald', 'sans serif';
+}
+.ant-menu-submenu {
+  background: rgba(112, 199, 131);
+  margin: 0.3rem;
+  border-radius: 6px;
+}
+.radio-buttons {
+  display: block;
+  height: '30px';
+  line-height: '30px';
+  font-family: 'Oswald', sans-serif;
+  font-weight: 400;
+  font-size: 1.0rem;
+  padding: 0.5rem 1rem;
+}
+#get-results-button {
+  position: absolute;
+  bottom: 2rem;
+  background-color: #ffffff;
+  width: 150px;
+  height: 50px;
+  border-radius: 6px;
+  font-family: 'Amatic SC', 'sans serif';
+  font-size: 2.5rem;
+  font-weight: 700;
+  
   display: flex;
-  flex-direction: column;
   justify-content: center;
   align-items: center;
-} */
+}
+#get-results-button:hover {
+  opacity: 0.7;
+}
